requirements.txt
PyQt5
aliyun-python-sdk-alidns
aliyun-python-sdk-core
requests
Pillow
pywin32; sys_platform == 'win32'
PyInstaller 